backend/requirements.txt
# Core web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Multipart file uploads
python-multipart==0.0.9

# Computer vision & numerics
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0

# Google AI
google-generativeai==0.7.2

# Google Cloud
google-cloud-storage==2.17.0
google-cloud-firestore==2.16.0

# Configuration
python-dotenv==1.0.1

# Type hints (Python < 3.11 compat)
typing-extensions==4.12.2
